+++ examples/ttl/series.go
@@ -123,8 +123,19 @@ func deleteExpired(ctx context.Context, c table.Client, prefix string, queue, ti
 				_ = res.Close()
 			}()
 
+			if !res.NextResultSet(ctx) {
+				if err := ctx.Err(); err != nil {
+					return err
+				}
+
+				if err := res.Err(); err != nil {
+					return err
+				}
+
+				empty = true
+				return nil
+			}
 			empty = true
-			res.NextResultSet(ctx)
 			for res.NextRow() {
 				empty = false
 				err = res.ScanNamed(